refactor: update product-related data classes to use BigDecimal for price and enhance response structures

+++ src/main/kotlin/org/shangahi/sellio_backend/api/dto/response/ProductCardResponse.kt
@@ -1,12 +1,13 @@
 package org.shangahi.sellio_backend.api.dto.response
 
+import java.math.BigDecimal
 import java.util.*
 
 data class ProductCardResponse(
     val id: UUID,
     val title: String,
     val description: String,
-    val minPrice: Double?,
+    val minPrice: BigDecimal?,
     val mainImageUrl: String?,
     val subCategoriesIds: List<UUID>,
     val isFavorite: Boolean
